+++ sdk/appconfiguration/app-configuration/src/internal/helpers.ts
@@ -69,27 +69,18 @@ export function checkAndFormatIfAndIfNoneMatch(
 }
 
 /**
- * Transforms the keys/labels parameters in the listConfigurationSettings and listRevisions
- * into the format the REST call will need.
+ * Transforms some of the key fields in ListConfigurationSettingsOptions and ListRevisionsOptions
+ * so they can be added to a request using AppConfigurationGetKeyValuesOptionalParams.
+ * - `options.acceptDateTime` is converted into an ISO string
+ * - `select` is populated with the proper field names from `options.fields`
+ * - keyFilter and labelFilter are moved to key and label, respectively.
  *
  * @internal
  * @hidden
  */
-export function formatWildcards(
+export function formatFiltersAndSelect(
   listConfigOptions: ListConfigurationSettingsOptions | ListRevisionsOptions
 ): Pick<AppConfigurationGetKeyValuesOptionalParams, "key" | "label" | "select" | "acceptDatetime"> {
-  let fieldsToGet: (keyof KeyValue)[] | undefined;
-
-  if (listConfigOptions.fields) {
-    fieldsToGet = listConfigOptions.fields.map((opt) => {
-      if (opt === "isReadOnly") {
-        return "locked";
-      }
-
-      return opt;
-    });
-  }
-
   let acceptDatetime: string | undefined = undefined;
 
   if (listConfigOptions.acceptDateTime) {
@@ -100,7 +91,7 @@ export function formatWildcards(
     key: listConfigOptions.keyFilter,
     label: listConfigOptions.labelFilter,
     acceptDatetime,
-    select: fieldsToGet
+    select: formatFieldsForSelect(listConfigOptions.fields)
   };
 }
 
@@ -215,4 +206,36 @@ function normalizeResponse<T extends HttpResponseField<any> & { eTag?: string }>
   delete newResponse.eTag;
 
   return newResponse;
+}
+
+/**
+ * Translates user-facing field names into their `select` equivalents (these can be 
+ * seen in the `KnownEnum5`)
+ *
+ * @param fieldNames fieldNames from users.
+ * @returns The field names translated into the `select` field equivalents.
+ *
+ * @hidden
+ */
+export function formatFieldsForSelect(
+  fieldNames: (keyof ConfigurationSetting)[] | undefined
+): string[] | undefined {
+  if (fieldNames == null) {
+    return undefined;
+  }
+
+  const mappedFieldNames = fieldNames.map((fn) => {
+    switch (fn) {
+      case "lastModified":
+        return "last_modified";
+      case "contentType":
+        return "content_type";
+      case "isReadOnly":
+        return "locked";
+      default:
+        return fn;
+    }
+  });
+
+  return mappedFieldNames;
 }