+++ sdk/appconfiguration/app-configuration/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.1.1 (Unreleased)
 
+- Fix issues with `select`ing fields to be returned from `listConfigurationSettings`, `listConfigurationRevisions`
+  and `getConfigurationSetting` where `last_modified` and `content_type` could not properly be passed in.
+  [PR #]()
 
 ## 1.1.0 (2020-07-07)